+++ tests/beaker_tests/cisco_interface/test_interface_svi.rb
@@ -1,5 +1,5 @@
 ###############################################################################
-# Copyright (c) 2014-2016 Cisco and/or its affiliates.
+# Copyright (c) 2014-2017 Cisco and/or its affiliates.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -46,10 +46,12 @@ tests[:default_mgmt] = {
   desc:           "1.1 Default 'mgmt'",
   title_pattern:  intf,
   manifest_props: {
-    svi_management: 'default'
+    description:    'Test default properties',
+    svi_management: 'default',
   },
   resource:       {
-    svi_management: 'false'
+    description:    'Test default properties',
+    svi_management: 'false',
   },
 }