+++ tests/beaker_tests/radius_global/radius_global_provider_nondefaults.rb
@@ -15,11 +15,11 @@
 ###############################################################################
 # TestCase Name:
 # -------------
-# SyslogServer-Provider-Defaults.rb
+# RadiusGlobal-Provider-NonDefaults.rb
 #
 # TestCase Prerequisites:
 # -----------------------
-# This is a syslog_server resource testcase for Puppet Agent on Nexus devices.
+# This is a radius_global resource testcase for Puppet Agent on Nexus devices.
 # The test case assumes the following prerequisites are already satisfied:
 # A. Populating the HOSTS configuration file with the agent and master
 # information.
@@ -29,12 +29,12 @@
 #
 # TestCase:
 # ---------
-# This is a syslog_server resource test that tests for default value for
-# 'ensure' attribute of a syslog_server resource.
+# This is a radius_global resource test that tests non-default attributes of
+# tacacs_global resource.
 #
 # There are 2 sections to the testcase: Setup, group of teststeps.
 # The 1st step is the Setup teststep that cleans up the switch state.
-# Steps 2-4 deal with cisco_snmp_group_resource and its
+# Steps 2+ deal with radius_global and its
 # verification using Puppet Agent and the switch running-config.
 #
 # The testcode checks for exit_codes from Puppet Agent, Vegas shell and
@@ -53,56 +53,71 @@
 
 # Require UtilityLib.rb and SnmpGroupLib.rb paths.
 require File.expand_path('../../lib/utilitylib.rb', __FILE__)
-require File.expand_path('../syslog_serverlib.rb', __FILE__)
+require File.expand_path('../radius_globallib.rb', __FILE__)
 
 result = 'PASS'
-testheader = 'syslog_server Resource :: All Attributes Defaults'
+testheader = 'radius_global Resource :: All Attributes Defaults'
 
-# @test_name [TestCase] Executes defaults testcase for syslog_server Resource.
+def cleanup
+  logger.info('Testcase Cleanup:')
+
+  command_config(agent, 'radius-server timeout 5')
+  command_config(agent, 'radius-server retransmit 1')
+  command_config(agent, 'no ip radius source-interface')
+
+  # To remove a configured key we have ot know the key value
+  on(agent, get_vshell_cmd('show running-config radius | include key'))
+  key = stdout.match('^radius-server key (\d+)\s+(.*)')
+  command_config(agent, "no radius-server key #{key[1]} #{key[2]}", "removing key #{key[2]}") if key
+end
+
+# @test_name [TestCase] Executes non-defaults testcase for radius_global Resource.
 test_name "TestCase :: #{testheader}" do
+  cleanup
+  teardown { cleanup }
+
   # @step [Step] Sets up switch for provider test.
   step 'TestStep :: Setup switch for provider' do
-    # Cleanup before test case start.
-    resource_absent_cleanup(agent, 'syslog_server')
-
     logger.info('Setup switch for provider')
   end
 
   # @step [Step] Requests manifest from the master server to the agent.
   step 'TestStep :: Get resource present manifest from master' do
     # Expected exit_code is 0 since this is a bash shell cmd.
-    on(master, SyslogServerLib.create_syslog_server_manifest_present)
+    on(master, RadiusGlobalLib.create_radius_global_manifest)
 
     # Expected exit_code is 2 since this is a puppet agent cmd with change.
     cmd_str = PUPPET_BINPATH + 'agent -t'
-    on(agent, cmd_str, acceptable_exit_codes: [2])
+    on(agent, cmd_str, acceptable_exit_codes: [0, 2])
 
     logger.info("Get resource present manifest from master :: #{result}")
   end
 
-  # @step [Step] Checks syslog_server resource on agent using resource cmd.
-  step 'TestStep :: Check syslog_server resource presence on agent' do
+  # @step [Step] Checks radius_global resource on agent using resource cmd.
+  step 'TestStep :: Check radius_global resource presence on agent' do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
-    cmd_str = PUPPET_BINPATH + 'resource syslog_server 1.2.3.4'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'severity_level' => '2' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'port' => '5555' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'vrf' => 'default' },
-                               false, self, logger)
-    end
+    cmd_str = PUPPET_BINPATH + 'resource radius_global default'
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output, { 'key' => add_quotes('44444444') },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'key_format' => '7' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'retransmit_count' => '4' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'source_interface' => "['loopback0']" },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'timeout' => '2' },
+                             false, self, logger)
 
-    logger.info("Check syslog_server resource presence on agent :: #{result}")
+    logger.info("Check radius_global resource presence on agent :: #{result}")
   end
 
   # @step [Step] Requests manifest from the master server to the agent.
-  step 'TestStep :: Get resource absent manifest from master' do
+  step 'TestStep :: Get resource present (with changes)manifest from master' do
     # Expected exit_code is 0 since this is a bash shell cmd.
-    on(master, SyslogServerLib.create_syslog_server_manifest_absent)
+    on(master, RadiusGlobalLib.create_radius_global_manifest_change)
 
     # Expected exit_code is 2 since this is a puppet agent cmd with change.
     cmd_str = PUPPET_BINPATH + 'agent -t'
@@ -111,23 +126,31 @@ test_name "TestCase :: #{testheader}" do
     logger.info("Get resource present manifest from master :: #{result}")
   end
 
-  # @step [Step] Checks syslog_server resource on agent using resource cmd.
-  step 'TestStep :: Check syslog_server resource presence on agent' do
+  # @step [Step] Checks radius_global resource on agent using resource cmd.
+  step 'TestStep :: Check radius_global resource presence on agent' do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
-    cmd_str = PUPPET_BINPATH + 'resource syslog_server 1.2.3.4'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               true, self, logger)
-    end
+    cmd_str = PUPPET_BINPATH + 'resource radius_global default'
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output, { 'key' => add_quotes('55555555') },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'key_format' => '7' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'retransmit_count' => '2' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'source_interface' => "['loopback1']" },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'timeout' => '2' },
+                             false, self, logger)
 
-    logger.info("Check syslog_server resource presence on agent :: #{result}")
+    logger.info("Check radius_global resource presence on agent :: #{result}")
   end
 
   # @step [Step] Requests manifest from the master server to the agent.
-  step 'TestStep :: Get resource present manifest from master' do
+  step 'TestStep :: Get resource present (with changes)manifest from master' do
     # Expected exit_code is 0 since this is a bash shell cmd.
-    on(master, SyslogServerLib.create_syslog_server_manifest_present_ipv6)
+    on(master, RadiusGlobalLib.create_radius_global_default)
 
     # Expected exit_code is 2 since this is a puppet agent cmd with change.
     cmd_str = PUPPET_BINPATH + 'agent -t'
@@ -136,48 +159,23 @@ test_name "TestCase :: #{testheader}" do
     logger.info("Get resource present manifest from master :: #{result}")
   end
 
-  # @step [Step] Checks syslog_server resource on agent using resource cmd.
-  step 'TestStep :: Check syslog_server resource presence on agent' do
+  # @step [Step] Checks radius_global resource on agent using resource cmd.
+  step 'TestStep :: Check radius_global resource presence on agent' do
     # Expected exit_code is 0 since this is a puppet resource cmd.
     # Flag is set to false to check for presence of RegExp pattern in stdout.
-    cmd_str = PUPPET_BINPATH + 'resource syslog_server 2003::3'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'severity_level' => '2' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'port' => '5555' },
-                               false, self, logger)
-      search_pattern_in_output(stdout, { 'vrf' => 'default' },
-                               false, self, logger)
-    end
+    cmd_str = PUPPET_BINPATH + 'resource radius_global default'
+    on(agent, cmd_str)
+    output = stdout
+    search_pattern_in_output(output, { 'key' => 'unset' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'retransmit_count' => '1' },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'source_interface' => "['unset']" },
+                             false, self, logger)
+    search_pattern_in_output(output, { 'timeout' => '5' },
+                             false, self, logger)
 
-    logger.info("Check syslog_server resource presence on agent :: #{result}")
-  end
-
-  # @step [Step] Requests manifest from the master server to the agent.
-  step 'TestStep :: Get resource absent manifest from master' do
-    # Expected exit_code is 0 since this is a bash shell cmd.
-    on(master, SyslogServerLib.create_syslog_server_manifest_absent_ipv6)
-
-    # Expected exit_code is 2 since this is a puppet agent cmd with change.
-    cmd_str = PUPPET_BINPATH + 'agent -t'
-    on(agent, cmd_str, acceptable_exit_codes: [2])
-
-    logger.info("Get resource present manifest from master :: #{result}")
-  end
-
-  # @step [Step] Checks syslog_server resource on agent using resource cmd.
-  step 'TestStep :: Check syslog_server resource presence on agent' do
-    # Expected exit_code is 0 since this is a puppet resource cmd.
-    # Flag is set to false to check for presence of RegExp pattern in stdout.
-    cmd_str = PUPPET_BINPATH + 'resource syslog_server 2003::3'
-    on(agent, cmd_str) do
-      search_pattern_in_output(stdout, { 'ensure' => 'present' },
-                               true, self, logger)
-    end
-
-    logger.info("Check syslog_server resource presence on agent :: #{result}")
+    logger.info("Check radius_global resource presence on agent :: #{result}")
   end
 
   # @raise [PassTest/FailTest] Raises PassTest/FailTest exception using result.