+++ CHANGELOG.md
@@ -2,6 +2,39 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [1.8.0] - 2017-12-12
+
+### New feature support
+#### Cisco Resources
+- `cisco_object_group` type and provider.
+- `cisco_object_group_entry` type and provider.
+
+### Added
+
+### Changed
+- `cisco_interface` Refactored to allow physical ethernet interfaces to be managed as ensurable resources.
+  - `ensure => absent` for physical interfaces will put the interface into a default state.
+  - `ensure => absent` for logical interfaces will cause them to be destroyed.
+
+- Extend `syslog_server` with attribute:
+ - `port`
+
+- Extend `syslog_settings` with attributes:
+ - `console`
+ - `monitor`
+ - `source_interface`
+ - `vrf`
+
+- Extend `radius_global` with attribute:
+ - `source_interface`
+
+- Extend `tacacs_global` with attribute:
+ - `source_interface`
+
+### Removed
+
+### Resolved Issues
+
 ## [1.7.0] - 2017-05-31
 
 ### New feature support
@@ -408,6 +441,7 @@ This version was never released.
 - Initial release of puppetlabs-ciscopuppet module, supporting Cisco NX-OS software release 7.0(3)I2(1) on Cisco Nexus switch platforms: N95xx, N93xx, N30xx and N31xx.
 - Please note: 0.9.0 is an EFT pre-release for a limited audience with access to NX-OS 7.0(3)I2(1). Additional code changes may occur in 0.9.x prior to the final 1.0.0 release.
 
+[1.8.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.7.0...v1.8.0
 [1.7.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.6.0...v1.7.0
 [1.6.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.5.0...v1.6.0
 [1.5.0]: https://github.com/cisco/cisco-network-puppet-module/compare/v1.4.1...v1.5.0